Frontend.Console(,Backend): support LN fee update
Update lightning channel fee rates when refreshing channels on
the frontend. Also, factor out a FeeUpdateRequired function
into the channel store so that checking whether a fee update is
required can be done separately to performing the update.

The "at your own risk" maybe needs some clarification to be
added later: if channel fee is not updated, the fundee might
decide to close the channel forcibly (e.g. with CPFP) in order
not to risk getting the channel-closing tx stuck in the
mempool.

Co-authored-by: Afshin Arani <afshinarani79@gmail.com>