Backend(LN): respond to peer's closing_signed if accepted
According to lightning spec, if closing_signed receiver agrees
with the fee, it should reply with a closing_signed with the
same fee_satoshis value, so other peer can create and broadcast
the closing transaction as well.

Before this commit geewallet wouldn't send anything back after
agreeing to the proposed fee which would cause the
closing_signed sender to get stuck waiting for a closing_signed
back (and Frontend.Console would hang).

* This commit also upgrades DotNetLightning.Kiss to version
  1.1.2-date20211122-1642-git-beb3109.
  (because this version includes a commit with the change
   needed to fix this bug, whose upstream PR is:
   https://github.com/joemphilips/DotNetLightning/pull/168)

Source:
https://github.com/lightningnetwork/lightning-rfc/blob/master/02-peer-protocol.md#requirements-6