scripts/make.fsx: refactoring, extracting new BuildSolution func
This refactoring gets us closer to how this script looks in the frontend
branch. We also add some type info to some arguments, to make this more
readable. And this opens the door to passing extra parameters to the
build tool (e.g. msbuild) such as /t:Restore, to explore a possible
restore target (mentioned as a TODO here[1]).

[1] e421a02d1b47f91bcb571106523e11794b6e4bb0