scripts/make.fsx: no need to call Mono in macOS (for nuget)
This is related to a recent refactoring[1] to reduce our dependence
on Mono in platforms where we don't care about its presence or not
(given that there's no reproducible-build system in macOS, unlike Linux*).

[1] d1afeb43d3e8a9be263d36197230c93bb7e75db4

* In Debian/Ubuntu we still want to be compatible with Mono, at least
until Microsoft makes sure that .NET5 gets in their default repositories
(as we don't want to rely on binaries created by Microsoft themselves if
we can).