GitHubActions: separate with steps instead of jobs
GitHubActions is not like GitLabCI, as it has the advantage of
being able to split a job in several steps. This way you can
know exactly which phase failed but you don't need to setup
everything from scratch again in each phase (in GitLabCI we
have workarounded this problem with docker images[1]).

This way we can mix all jobs in just one workflow (CI), in
which each job has the simply the name of the OS it runs on.
Then we have a 2nd workflow that is the only Linux-based one
(cause Linux builds we already have with GitLabCI) just for
publishing the app in the Snap store.

The only disadvantage this has is that we have to clean up
binaries before compiling in different mode (DEBUG vs STRICT
vs RELEASE) because there seems to be an msbuild bug (it
considers the build done after adding a new define, even
if the previous build wasn't run with that define...) which
I added as a FIXME in the CI.yml file.

[1] 2bfc26c6658051144726837a8e12cedd19d7970e