CommonBuildProps: improve potential build error msg
If the targets file is not found, the build would error
with something like:

GWallet.Backend.fsproj : error MSB4057: The target
"Build" does not exist in the project.

But with this change, the error message is much more
understandable:

GWallet.Backend.fsproj(43,3): error MSB4019: The
imported project
"/usr/lib/mono/xbuild/Microsoft/VisualStudio/v16.0/FSharp/Microsoft.FSharp.Targets"
was not found. Confirm that the expression in the Import declaration
"/usr/lib/mono/xbuild/Microsoft/VisualStudio/v16.0/FSharp/Microsoft.FSharp.Targets"
is correct, and that the file exists on disk.