scripts: Rework sanity-check to parse solution files
The part of the sanitycheck target which checks that
nuget packages are in sync (same versions, no unused
packages, missing packages, ...) started to fail when
being imported in the frontend branch; case in point
being it detecting packages not being restored (e.g.
OpenTK from the Frontend.XF.GTK project, for the
Windows build agent).

Even though the GTK project can be built in Windows,
we prefer to only build it in the Linux lane. Anyway,
a similar problem might arise for, let's say, nuget
packages of the iOS frontend in the Linux build agent.

The reason for this to happen is that the algorithm
that scans for packages starts with the solution file
but it didn't parse it, and instead it just looked for
project files in directories underneath the directory
where the solution file is. This way, nuget packages
will be detected for projects that don't even get
built for certain OS, and therefore don't get restored.

With this commit we change the nuget sanity check to
really parse solution files via the "Microsoft.Build"
API supplied by the nuget package of the same name. It
was a bit tricky to get this to work because the script
itself has to download a nuget package, which shouldn't
be restored in the same folder where all the other nuget
packages are restored (to bypass the excess-packages
detection, because no project really references it).
This requires an updated 'fsx' submodule because we
bring some bugfixes from there to be able to have this
work in linux-stockmono build agents.