CI(snap): fix snap generation to include GTK frontend, not console
This got broken when somehow GitHub decided to not install fsharp's
mono package anymore in their default VM. The first bad consequence
of this is that Linux builds were failing (so snap generation of
console frontend, the snap called "gwallet"; not "geewallet", failed).
We fixed that problem in the master branch with the following workaround:
9d76c300687450c4d737338930f7b2559a80eaa9

However, this meant that Linux builds on Github were not using anymore
msbuild, but xbuild along with an older F# (4.0). This is a problem in
the frontend branch because xbuild cannot build .NETStandard projects,
and our Xamarin.Forms main UI project is a .NETStandard library, so it
was not being built. The consequence of this is that the last version
of our 'geewallet' snap package, 0.4.254.0, started failing to launch
with the following error:

/snap/geewallet/34/bin/desktop-launch: line 611: /snap/geewallet/34/bin/geewallet-gtk: No such file or directory

In fact, checking what is in that directory can let us see that the gtk
frontend is not there anymore, but the console one:

$ ls /snap/geewallet/34/bin/
desktop-launch  geewallet-console

To fix this, we need to explicitly install fsharp's deb package from
Microsoft's mono-project.com site, which brings v5.0 (instead of v4.0)
which is compatible with the latest msbuild, so that we don't need to
fallback to xbuild anymore, and the GTK frontend gets built again.