Revert "Backend,Tests: Update NBitcoin to v6.0.17"
This reverts commits a16ef649386e59f5f3c3da25933ba5ed6e198886
and 3a1587263c44579248ca8370e0eeb45698a88fc5 because the LN
branches were using a previous version of NBitcoin, and these
branches used DotNetLightning releases which are not compatible
with this new version of NBitcoin, since it breaks backwards
compatibility (API and ABI).

We will revert this revert in the last commits of this branch,
where we have more control over what NBitcoin version DNL (or
DNL.Kiss) can use.