scripts: move snap release from GitHubCI to GitLabCI
The reason to use SNAPCRAFT_LOGIN_FILE instead
of SNAPCRAFT_LOGIN (like in Github) is in the docs:

From https://docs.gitlab.com/ee/ci/variables/#cicd-variable-types :

...
File type variables:

Consist of a key, value and file.
Are made available in jobs as environment variables, with
The CI/CD variable key as the environment variable name.
The CI/CD variable value saved to a temporary file.
The path to the temporary file as the environment variable value.
...

And the reason for using "needs:" instead of "dependencies:"
is because the job depends on artifacts of a job in the same
stage. More info: https://gitlab.com/gitlab-org/gitlab/-/issues/30632

Co-authored-by: Andres G. Aragoneses <knocte@gmail.com>