snap: upload pkg inside docker too
Because snapcraft needs to be run also to upload
the snap package. To do this, we need to pass the
environment variables needed from the host of the
container (and even the empty ones, that's why we
remove the -u modifier, to fix a possible "unbound
variable" problem, see [1] for more info).

I found out about this when I tried to release the
v0.3.256.0 tag, and it failed when uploading the
snap [2], because the upload process also needs to
run snapcraft, and this was not installed in this
gitlabCI build step (and installing it wouldn't work
because inside GitLabCI it can only be run inside
a sub-docker container). This is also the reason why
we merge both snap_ build steps into one, otherwise
we would need to insert the snap file into yet
another docker container in order to push it later,
but the snap file is already generated in a docker
container, so why not push it from that point? This
way it's much simpler.

[1] https://unix.stackexchange.com/a/463037/56844
[2] https://gitlab.com/nblockchain/geewallet/-/pipelines/416574780