scripts: check GitHub's mirror, not canonical mirror
CI in forks/MRs got broken when we introduced GitLab's
snap generation [1] because it assumed that env var
SNAPCRAFT_LOGIN_FILE should be there (as it is, in
our canonical repo, but it doesn't need to be in forks,
so CI pipelines of MRs kept failing). I tried to fix
this here[2] and here[3] but it wasn't enough, because
a branch in someone's GitLab fork may not be in the
canonical GitHub mirror of geewallet.

So this time we check local remotes to find the GitHub
mirror of the fork we're running on, not the canonical
mirror. This will force MR submitters to setup a GitHub
mirror from their GitLab fork but it's a good thing, to
make sure they don't break macOS & windows builds (as
we don't have those in GitLabCI).

[1] 8b87b15bd7f69860fff2c0b77ea2971f4b3f90cf
[2] ef9c07b64389f48c5fef5c50a29f57bfe01c4c88
[3] b9a499a64ad4b19962b8366def0c18d32d110127