End2End: process dispose should happen in order
If we dispose out of order (e.g bitcoind before electrs), the
dependent process might die before we dispose it and cause crash
when we try to dispose it.

Order was already correct in the TearDown function but since the
signature wasn't completely type enforced arguments were passed
out of order.