Frontend.XF(WelcomePage): clarify form is not an account signup/login
Google(Play) staff is asking us for login credentials to test (again) the
app, and we are insisting to them that this welcome page is not a log-in
or sign-up page of any kind. By adding this extra information we hope that
they stop nagging us to provide something that can generate themselves by
typing any random data.