Backend(LN),E2E,FE.Console(L2): anchor channels
Supporting anchor channels add the possiblity for both parties
(funder&fundee) to use CPFP for speeding up the confirmation in
the case of force-closing a channel (previously CPFP support
only was allowed with mutual-close).

NOTE: the change in ChainWatcher.fs is to only broadcast penalty
if/when the spending tx is confirmed (because after this commit,
Geewallet's penalty tx tries to spend to_remote output as well
and that output is time-locked by 1 block, so we need to wait
for at least 1 confirmation before broadcasting the penalty tx;
otherwise, "no-BIP68-final" error would be thrown when
broadcasting the penalty tx before the 1 block confirmation).