Frontend.*,Backend.NetStandard: sync with System.Memory upgrade
Last merge included an upgrade (from master branch) of System.Memory,
so we need this to sync with the frontend projects.

Note: the Backend.NetStandard project didn't have a direct dependency
on System.Memory, but it was a sub-dependency of System.IO.Pipelines;
then we need to add a direct dependency like it's done in this commit,
in order to not update System.IO.Pipelines.