Frontend.XF: convert some dynamic UI code to static XAML
There was no reason to draw QR codes in 2 different ways like
it was happening before this commit:
- ReceivePage: added programmatically via F# code behind.
- PairingFromPage: using zx/zxcm tags statically inside XAML.

In general, it's better to use static XAML when there's no need
for dynamic content, so that it could potentially be edited with
a WYSIWYG editor (like VS has/had for Xamarin).

Also, having different ways of showing QR codes exposed a bug
which was kinda hard to catch: when upgrading ZXing.Net.Xamarin
libs, a change of an assembly name[1] was causing an issue only on
the PairingFromPage in offline-mode (and given that offline-mode
is harder to debug with an emulator, it was tricky to figure out
what was going on). If we had had XAML-based QR codes in the
ReceivePage too (like after this commit) then we would have caught
the problem much earlier and in online-mode too.

[1] If you're curious, the assembly name change that happened in
the ZXing.Net.Xamarin upgrade was from "zxing.portable.dll" to
"zxing.dll".