Backend(FaultTolerantParallelClient): improve a potentially misleading ex
If the minimumNumberOfResponsesRequired for an AverageBetweenResponses type
of request was higher (e.g. 2) than the number of funcs passed in (e.g. 1),
a very confusing ArgumentException would be thrown, whose message would be
"number of funcs must be higher than zero" (and it was not 0, it was 1!).

With this commit we fix it to throw an ArgException earlier that would have
the proper non-confusing message. In this commit we also improve another
ArgumentException's message that we were already throwing, but now including
the particular counts being compared.