Frontend.XF(.*),Backend: upd ZXing.Net.Xamarin & DNE to last versions
* ZXing.Net.Xamarin: to bring this change [1] (it didn't make much sense
to map the net461 profile to Linux, because a) we might run Linux+GTK
clients with .NET6.0 soon; b) you could run a net4xx frontend on Windows).
* DotNetEssentials: 1.6.1x version syncs with Xamarin.Essentials 1.6.1
(and this version is now a dependency of ZXing.Net.Xamarin v3.x.x anyway).

NOTE: this is trying again to upgrade but with a new version[2], given
that the previous version update[3] failed to build [4] so it had
to be reverted [5].

[1] https://github.com/nblockchain/ZXing.Net.Xamarin/commit/4a604917982af501cb5b64100794eefbb5417018
[2] https://github.com/nblockchain/ZXing.Net.Xamarin/commit/462686e58c743d2d6c4230b4e6f92fc9053616f7
[3] https://gitlab.com/nblockchain/geewallet/-/commit/29e2d3a08de24e464656a2e70864c91001c81a19
[4] https://gitlab.com/nblockchain/geewallet
[5] https://gitlab.com/nblockchain/geewallet/-/commit/e2f1778703b83130e582e99ccb9d41f08060f499