GitHubCI: add a linux-vanilla lane
This is necessary because GitHubCI adds a lot of
preinstalled software by default in their agents;
so we want an agent without this.

BTW, the need for the ownership workaround was
to prevent this error below:

```
checking for a working F# REPL... not found
checking for fsx... not found
Populating sub-fsx module...
fatal: detected dubious ownership in repository at '/__w/geewallet/geewallet'
To add an exception for this directory, call:

	git config --global --add safe.directory /__w/geewallet/geewallet
./configure.sh: line 40: ./configure.sh: No such file or directory
./configure.sh: line 52: scripts/build.config: No such file or directory
Error: Process completed with exit code 1.
```