ReadMe.md,CONTRIBUTING.md: phase out GitLabCI/GitLab
GitLab has historically aligned better with our opensource values
and that's why we were using it predominantly over GitHub, however,
there are many reasons why we're moving out of it.

At first, we started using GitLabCI mainly because it was better
than the competitors out there like TravisCI, mainly because it
was integrated with the git service (you push and get a CI job
automatically, no need to hookup/config integrations). We mainly
cared about Linux so we didn't mind much that GitLab.com didn't
offer Windows/macOS buildagents (we could always plug them from
our cloud provider if we needed to).

Then GitHub came along and offered GitHubActions, a free (for
opensource) CI service that offered all 3 OSs: Linux, macOS and
Windows. So then we added GitHubCI jobs for Windows and macOS,
and left the GitLabCI jobs for Linux. We still preferred the
GitLabCI's Linux jobs over the GitHubActions ones because the
latter came with preinstalled software that we didn't control
(e.g. the ubuntu VMs spawned for these CI jobs in GitHubActions
included the latest versions of mono, and we cared about testing
with old versions of mono too, so we prefer to update mono on
our own).

Then GitLab.com started restricting the free number of CI minutes
that accounts get for private repositores (we thought: sure, we
are opensource so this doesn't affect us that much).

Then GitLab.com started requiring credit card details for people
that wanted to create accounts and run CI jobs (even on opensource
public repos). This annoyed us because we didn't want to ask our
contributors to trust GitLab with their credit card details, even
if they would just use these for "verification" (spammer protection
I guess), not purchases.

Then GitLab.com started limiting CI minutes even for opensource
public repos, and then we hit the limit. Ok, now what? Well, we
have now discovered a way to run vanilla ubuntu versions in
GitHubCI (via the `container:` attribute), so we can now switch
over completely to GitHubActions.

On top of these, we never liked the way GitLabCI works to maintain
state between job steps: in GitHubActions this is automatic, in
GitLabCI you need to setup image registries or pass artifacts from
one step to the next, etc.

Good bye GitLab and thanks for all the fish. We still like that
you are opensource and GitHub is not, but your freemium deployment
(as a company, not product) has killed our opensource use-case.