GitHubCI(Linux+GTK): retry 'make' 3 times (flakey nuget)
Previous commit fixed indeed the 'sanitycheck' of some builds
affecting Linux+GTK, however, one of the builds failed at the
'make' step. By looking at the log, you can see the compiler
error hints at the Xamarin.Forms dependency not being present,
so this means that nuget restore didn't work properly.

If a nuget restore operation doesn't work, it should fail in
that process and stop; not sure why it didn't fail this way.
But it's not worth to investigate because:

1) In the past, I implemented[1] a solution to have the nuget
restore operation be part of the normal build, kind of like
a PreBuild step via MSBuild. The code that allowed this is
kinda horrible and don't want to debug at the moment, because...

2) We're going to move to .NET6 soon anyway, which means not
using nuget anymore (in favour of the "dotnet restore" command).

Given the above, I'm using here temporarily a very hacky solution:
retry the 'make' step 3 times. And I also doing it in a very ugly
way because I'm very wary of using random people's githubActions,
such as what is the existing/recommended solution for this in
the GitHubActions marketplace [2].

[1] 9804b653a63266c8560c7e8143f9ec644f05a334
[2] https://github.com/marketplace/actions/retry-step