DNL.Kiss v1.1.2-date20210412-1027-git-6062c26 upd
This version supports closing (& force-closing) channels (comes
from this branch [1]), but we don't make use of this feature yet
(upcoming commit next to this one will).  NOTE: the
ChannelMarshalling unit test had to be expanded to catch
possible bugs in PerCommitmentSecretSore deserialization.

The reason why we add System.Runtime.CompilerServices.Unsafe to
Frontend.Console is because after this upgrade, if we didn't, we
would get FileNotFoundExceptions when running it (integration
tests that run --update-servers flag), and not even
BindingRedirects or upgrading to newer versions would fix it.
And even after this reference was added there, the same
exception would happen on Windows OS, so it had to be upgraded
to 4.5.3 including a binding redirect (as adviced here: [2]).
As I was upgrading the version to 4.5.3, then I also upgraded it
everywhere in the solution.

[1] https://github.com/nblockchain/DotNetLightning.Kiss/commits/geewalletLightningMilestone3
[2] https://stackoverflow.com/a/62770487/544947

Co-authored-by: Andrew Cann <shum@canndrew.org>