Backend.Tests.End2End: swap electrumX with electrs
We need to replace electrum-x with electrs because we're being
affected by a bug in it which seems to be very similar (if not
the same) as this:

https://github.com/kyuupichan/electrumx/issues/630

Maybe this bug was fixed there but not in spesmilo's version?
Anyway electrs is more efficient software and we like Rust more
than Python in case we need to fix anything in it.

NOTE: we now mine blocks to dummy address, because to
geewallet's address caused electrs to hit a limit where it can
no longer answer requests pertaining to that address.  Geewallet
is not able to use funds from coinbase outputs and we only mine
blocks to it for the sake of confirming other transactions.  As
such we can replace all instances of mining blocks to geewallet
with mining blocks to a throw-away address instead.  This is
related to this other recent commit:
0cfd232f3eddf642e21850321fb0420572a89781