scripts: deduplicate InfraLib in favour of fsx's
The InfraLib/ folder inside scripts was just an old version of
fsx's InfraLib, and given that a) we just fixed an important
hang in the latter; b) we already had a submodule for it; then
we're switching to just use the submodule directly.

This has two consequences:
1) The submodule has to always be populated, so we changed the
configure script to always do it.
2) InfraLib's Process API has been overhauled so there's a lot
of changes related to that in this commit.