scripts/snap_release.fsx: move login check later
Even if we already had a SNAPCRAFT_LOGIN check that took in
account forks, there's an edge case that made this not work:
when a CI is spawned from a PR event.  This way, it seems the
repo is nblockchain's but the repo secrets come from the fork
o_O ! Strange, I know.

Instead of investigating more I'm just going to add a
workaround, so we move the git tag check earlier, so this edge
case won't happen (as there will be no PRs that are git tags at
the same time).

This bug had happened in the CI[1] of this PR[2].

[1] https://github.com/nblockchain/geewallet/actions/runs/3385674635/jobs/5624757832
[2] https://github.com/nblockchain/geewallet/pull/180