GitHubCI: remove xbuild from name of some lanes
It turns out that when I created these jobs, the github VMs
didn't have msbuild installed by default, so I suffixed their
names with "xbuild".  This seems to have changed now, so to stop
getting confused, I simply remove that suffix, and I replace
"msbuild" suffix with "newmono" to be consistent with the naming
of the vanilla jobs.