scripts/fsx: upgrade for .NET6 compatibility
For this to work, we've had to do some changes on how F# scripts
are run, using LEGACY_FRAMEWORK when legacy tools (such as
fsharpi) are used.

PS: don't ask me why we need the +x for the "run" makefile
target or why it wasn't needed before, but this prevents a
"Permission denied" error that was popping up after these
changes.