GitHubCI,scripts,configure.*: use Fsdk nuget
InfraLib has been renamed to Fsdk in fsx and packaged as a nuget
lib so that we don't need to populate the git submodule for
compiling with dotnet6 anymore.

Reason for disallowing warnings now when calling Process.Execute
is because by using Fsdk nuget package we workaround the FSI
bug[1] that emits these warnings. (We allowed these warnings
in [2]).

[1] https://github.com/dotnet/fsharp/issues/14216
[2] f5f15c4d9a70115b2325d4d1de63346f462e119f