scripts/snap_build.sh: don't use dotnet6 (hack)
Since we merged this PR[1] to default to dotnet6 build, the
snap package is broken because it's architected with mono in
mind. Apparently GitHubCI's VMs don't have dotnet installed
via apt so we cannot purge it with this command, so we use
this hack for now until we find a better solution.

[1] 7e6bc84b912eff8f4b8513f35a48fb15b6f4cad4