GitHubCI,scripts: more fault-tolerant apt install
This morning GitHubCI is taking sometimes 30minutes to run some
jobs that used to take 10minutes. The dotnet6 job in other
branches is even taking more than 1hour! And the bottleneck
seems to be when downloading and installing apt packages.

Not sure if it's a temporary glitch today or GitHub has decided
to start throttling this, but it's making apt failures happen
more often, so we need to develop a smart retry mechanism,
which I have encoded in these two scripts (which are also smart
in the sense that you don't need to use sudo anymore in the
GitHub YML file, and no need to apt update before either, plus
it sets the DEBIAN_FRONTEND=noninteractive env var just for
you).