scripts/fsx: update submodule to latest master
The change in configure.sh is needed because with the non-legacy
NuGet.config, fsx would restore packages in the modern way (as
in lowercase package name and version as a subfolder, i.e.
/packagename/version/ , as opposed to /PackageName.Version/),
which the sanitycheck.fsx script didn't understand for the
legacy side of it.

And we only noticed the problem above with this submodule update
because we have added nuget packages recently[1] in fsx.

[1] https://github.com/nblockchain/fsx/commit/2550578b80e5f7a0bfead586ce2dd31511299cd0