GitHubCI: rework stock dotnet job(s)
* Rename the vanilla--dotnet6 job to vanilla--stockdotnet6 to be
consistent with how other jobs are named (stock means that the
package comes from Ubuntu without installing external sources).

* Create new vanilla--stockdotnet6-and-stockmono job that
installs both dotnet6 and mono, from Ubuntu repos.  A scenario
in which master branch would just default to use dotnet6 (not
mono) but which:

1) Should be tested anyway in CI, just in case there are
regressions.

2) Might be used soon in the 'frontend' branch to generate the
binaries for Frontend.XF.Gtk.  (Right now the only way to
generate these binaries is via msbuild, which is a package not
provided in Ubuntu, as in, the jobs where these binaries are
generated are not "stock" style, which is not a good thing
because it makes us depend on Microsoft).