GitHubCI,scripts: upload snap pkg as artifact
In the stable branch, we used to do this, but mainly because we
needed to pass the artifact from one job to the next.  In master
branch there is a single job because we merged them into one
(because there was no point in separating them) and then we
didn't need the artifact.  However, there's no harm in uploading
the artifact because by being opensource&public repo we don't
get charged for storing it.

Also there's no point in changing the retention policy because
it defaults to 90days and 90days is the maximum anyway.  If this
was different, we could change bump.fsx to increase the
retention for releases (with git tags) and decrease it for
commits in between releases, but there's no real benefit with
the current policies.