scripts/make: build XF frontend when dotnet exists
Before this commit, we were assuming that legacy tools
were not present if dotnet was present. Now, if dotnet
is present, and legacy tools also are (so that the XF
frontend can be built), then make will build the frontend.

Obj path in some projects had to be changed like it was
done in [1] because building same project once with dotnet
and once with MSBuild causes problems (it happens in macOS
when both mono and dotnet exist).

[1] 6b3c64ed493e9481560e96e1858a938aaa1b56e7