Frontend.XF: stop hiding 0 DAI accounts
When MakerDAO made DAI be multi-asset instead of single-asset,
they renamed the old DAI to SAI and they made "DAI" be the new
multi-asset token. I was not happy about this approach and so
I stopped hiding this coin when the user had a zero balance of
it (if it had a positive balance, I would of course not hide
their funds; but hiding it for zero meant that it wasn't
"default" anymore).

However:

* The Console frontend still shows today DAI when zero balance,
which is inconsistent. Not sure if it's because I forgot to do
the change there or because I've changed my mind and I removed
the hiding there to begin with.

* Multi-asset DAI is now pervasive, and being rebelious about
it doesn't make much sense.

So for the sake of consistency (with the Console frontend) and
to still make this wallet have a stablecoin by default (which,
despite being multi-asset, is still the least worse among other
alternatives), let's change this back to show it again for zero
balances.