scripts/snap_build.sh: remove --destructive-mode
As far as I have gathered this morning, snapcraft has 3 modes
of operation: 1) Host, 2) Multi-pass, 3) LXD.

According to `snapcraft build --help` the --destructive-mode
flag implies using the 1) Host above.

Since we've had mild issues [1] when upgrading our snap package
generation from Ubuntu 20.04 to 22.04, and since apparently LXD
is the default mode for snapcraft when using 22.04 (or core22,
not sure), I tested removing this flag to see what happens. But
then got this error:

```
You need 'multipass' set-up to build snaps: https://multipass.run.
```

So I guess that given we're pinning a snapcraft version (a
channel actually) that is still kinda old (in the sense that
LXD is still not the default in this case). Therefore, let's
force LXD usage by using --use-lxd this time.

BTW we need sudo and LXD initialization because otherwise we
would get an LXD socket error (see [2] and [3]).

[1] https://forum.snapcraft.io/t/gtk-related-issues-when-upgrading-snap-pkg-generation-from-20-04-to-22-04/36537
[2] https://github.com/samuelmeuli/action-snapcraft/issues/11
[3] https://snapcraft.io/docs/build-providers