Backend: update Nethereum to latest version
This commit updates Nethereum to latest version to support
extracting token transfer amount from the raw tx created when
sending (hot-storage) or signing (cold-storage).

It's noteworthy to say that Web3 no longer accepts a timeout
as an argument.

Nethereum now supports different types of transactions like
those that support EIP-1559, at this moment there's no
clear benefit for migrating especially given the fact that
it will causes incompatiblity with ETC so we will continue
to use Legacy types for now.