Backend.*: fix assembly version attribs
Somehow, even if CommonAssemblyInfo.fs is included in
Backend.NetStandard project, it was not being applied (as the
recent snap testing commit [1] revealed, when it calls
geewallet with the --version flag), so we now use MSBuild
attributes instead, which work and the best proof that they do,
is that we've had to change a LOC in Backend.Tests for them not
to break after this change.

NOTE: this commit was decided to be backported from master to
stable branch, cherry-picking [2], where Backend.NetStandard
project doesn't exist, but Backend project does.

This is the 2nd part of the bugfix for bug 242.

[1] ce93f84c23b8ce8a682c56a3df3ebca2e4df528a
[2] 2edb96ead1abc1ab2450c5e197cfbc50b2cec70a