Backend: morph exception wrapping at balance chk
This InconsistentResultFromDifferentServersOfSameCurrency
exception type was introduced in [1], but after implementing an
improvement[2] over the exception message and type of another
problem, I realised that the new exception thrown in [2]
wouldn't be covered by the exception type mentioned above, so we
need it to be more generic (to receive any exception as
innerException), so let's do that and rename it.

This commit in conjunction with [2] don't really fix issue
250 [3], but it might mitigate it, and anyway thanks to [1] if
it happens again, we will receive a different exception, with
different exception message as well (more helpful), so let's
just close [3] now.

Closes https://github.com/nblockchain/geewallet/issues/250

[1] 8fa47e0e2bd1b325761c7bdc9f8e32a5955b0f5c
[2] 5eca67170579b44b9598e57656fbd3118b9e14dd
[3] https://github.com/nblockchain/geewallet/issues/250