Backend: informative telemetry for SAI
We're interested in phasing out the SAI cryptocurrency support,
but before doing so, it would be good to know if someone is
still using it. Sending a Sentry warning when a balance is
higher than zero should be able to tell us (if in 3 months we
receive no warnings, I guess we can go ahead with dropping it).