Backend,Frontend.Console: rename JSON property
This way, if a version of geewallet that is too old to have
native segwit support (or disabled by default) tries to
deserialize a transaction proposal, it will error out
complaining about version mismatch instead of crashing with a
weird error.

This commit also tweaks the version comparison, to make sure
the VersionMismatch exception is actually raised, and caught
by the Frontend (Console, for now).