Backend: migrate ReadOnly accounts to NativeSegwit
Their public addresses would then start showing their native
segwit address (that starts with "bc1") instead of the nested
segwit one. This way they don't need to re-scan the QR code
from the coldstorage device. If we don't do this, if you tried
to sign-off the transaction on your native-segwit-ready cold
storage device, you would get this error:

```
Introduce a file name to load the unsigned transaction: testCold.json
Error: The transaction doesn't correspond to any of the accounts in the wallet.

Press any key to continue...
```

In case you're wondering if there's any risk that this causes
irrecoverable funds when you use this with a coldstorage
device that you haven't upgraded yet to NativeSegwit: this
could not happen since the JSON format is different, and then
deserialization would fail, advising the user to upgrade.