Frontend.Console: avoid always showing currency
Recent progress-showing feature introduced in [1] was showing
the currency ticker also when querying the balance of a single
account (something which is used in other scenarios, not in the
onboarding print of balances that you get at first when you
launch the Frontend.Console). With this showProgress param we
now only print it when it's useful.

[1] b72bb8659a96cac3f465ac679a202467eb54fdab