Backend/Config: fix ETH-token readonly removal
Removal of read-only accounts for ETH tokens was not working
because the code to copy public-key files (at the time written
probably so that people with only SAI accounts got DAI accounts
automatically, due to the SAI->DAI migration) was recreating
the files again. Now, we check if the directory was already
there, in which case we don't copy the files (so that this
token-creation code only happens when the folder for the token
didn't exist yet).