FE.XF.GTK: open console app with --console arg
This commit allows for XF.GTK to open/call console version when
passed the "--console" argument.

For Frontend.Console to be callable/referenceable from
Frontend.XF.GTK, it needs to be built in .netstandard2.0 instead
of net6.0, so this commit does that. .NETStandard build happens
when either a flag is passed (for stockdotnet6+stockmono lane)
or when we're building on mono (for other lanes building
XF.GTK).