configure.{sh,fsx}: fix bad order of tool findings
When porting stable's branch commit [1] to master branch [2] we
missed the fact that our configure.fsx script in the master
branch checks for `dotnet` executable before the OS match block
so the " found" string was not being printed first.

This fixes it with a flag so that if someone calls the .fsx file
by hand then the " found" string doesn't show up in that case.

[1] ae9f2a6a8ab94b2734a15375cb8a3ca8b7c14964
[2] 0b23222b837b54623cb0c1a064781a3692cfe06a