scripts/make.fsx: run unit tests in DEBUG mode
Unit tests were already being run in DEBUG mode for .NET legacy
pipelines, but for new dotnet pipelines it was running in
Release due to the absence of the --config flag.

This was noticed when I introduced some code with an F# warning
(which should fail the build when trying to compile in RELEASE,
but I noticed that it failed the build on dotnet lanes at the
unit tests CI step).